fix: ah it was... reaction

+++ controllers/reaction_controller.ts
@@ -38,10 +38,7 @@ const reactionController = {
         `
 
         const result = await c.orm.getRepository(Reaction).createQueryBuilder().where(query).getOne()
-        if (result) {
-          if (result.deletedAt) return c.res.failed()
-          await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
-        }
+        if (result) await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
         else await c.orm.getRepository(Reaction).insert({
           ip: c.req.ip,
           type: c.req.body['type'],
@@ -70,10 +67,8 @@ const reactionController = {
         `
 
         const result = await c.orm.getRepository(Reaction).createQueryBuilder().where(query).getOne()
-        if (result) {
-          if (result.deletedAt) return c.res.failed()
-          await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
-        } else await c.orm.getRepository(Reaction).insert({
+        if (result) await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
+        else await c.orm.getRepository(Reaction).insert({
           ip: c.req.ip,
           type: c.req.body['type'],
           reply: { id: c.req.body['replyId'] },
@@ -101,10 +96,7 @@ const reactionController = {
         `
 
         const result = await c.orm.getRepository(Reaction).createQueryBuilder().where(query).getOne()
-        if (result) {
-          if (result.deletedAt) return c.res.failed()
-          await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
-        }
+        if (result) await c.orm.getRepository(Reaction).createQueryBuilder().where(`id = ${result.id}`).delete().execute()
         else await c.orm.getRepository(Reaction).insert({
           ip: c.req.ip,
           type: c.req.body['type'],
